requirements.txt
urllib3
google-api-python-client